ISSUE44-update-get-schedule
