+++ src/controllers/user.js
@@ -10,8 +10,6 @@ const DataFormatError = require('../errors/DataFormatError');
 const {
   validateJoinSchema,
   validateScheduleSchema,
-  validateYYYYMMDDDateSchema,
-  validateYYYYMMDateSchema,
 } = require('../utils/validators');
 const { UserNotFoundError } = require('../errors');
 
@@ -54,40 +52,19 @@ async function putUserProfile(req, res, next) {
   }
 }
 
-async function getUserPersonalMonthSchedule(req, res, next) {
+async function getUserPersonalSchedule(req, res, next) {
   try {
-    const { error } = validateYYYYMMDateSchema(req.query);
+    const { error } = validateScheduleSchema(req.body);
     if (error) return next(new DataFormatError());
 
-    const { date: dateString } = req.query;
     const user = await User.findOne({ where: { nickname: req.nickname } });
     if (!user) {
       return next(new UserNotFoundError());
     }
 
-    const start = moment.utc(dateString, 'YYYY-MM').startOf('month').toDate();
-    const end = moment.utc(start).endOf('month').toDate();
-    const schedule = await PersonalSchedule.getSchedule(user.userId, start, end);
-    if (schedule === null) throw new ApiError();
-    return res.status(200).json(schedule);
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function getUserPersonalDaySchedule(req, res, next) {
-  try {
-    const { error } = validateYYYYMMDDDateSchema(req.query);
-    if (error) return next(new DataFormatError());
-
-    const { date: dateString } = req.query;
-    const user = await User.findOne({ where: { nickname: req.nickname } });
-    if (!user) {
-      return next(new UserNotFoundError());
-    }
-
-    const start = moment.utc(dateString, 'YYYY-MM-DD').startOf('day').toDate();
-    const end = moment.utc(start).endOf('day').toDate();
+    const { startDateTime, endDateTime } = req.body;
+    const start = moment.utc(startDateTime).toDate();
+    const end = moment.utc(endDateTime).toDate();
     const schedule = await PersonalSchedule.getSchedule(user.userId, start, end);
     if (schedule === null) throw new ApiError();
     return res.status(200).json(schedule);
@@ -116,7 +93,6 @@ async function putUserSchedule(req, res, next) {
 module.exports = {
   getUserProfile,
   putUserProfile,
-  getUserPersonalMonthSchedule,
-  getUserPersonalDaySchedule,
+  getUserPersonalSchedule,
   putUserSchedule,
 };
