+++ src/utils/validators.js
@@ -69,6 +69,7 @@ const groupScheduleSchema = Joi.object({
 });
 
 const userScheduleSchema = Joi.object({
+  userId: Joi.number(),
   title: Joi.string().max(45).required(),
   content: Joi.string(),
   startDateTime: Joi.date(),
